Level Editor: Simple top-down style play test mode added.
TileMap:
Can check collision for red and blue rectangles.

Images2D:
A new constructor that can specify if the image should be loaded as a
resource, from inside jar file. Image2D(path, useRecourceLoading)

-----Bug fixes----

Rectangle:
There was a bug when checking collision with bottom left corner of an
other rectangle.